+++ codex-rs/tui/src/lib.rs
@@ -66,7 +66,6 @@ mod streaming;
 mod text_formatting;
 mod tui;
 mod ui_consts;
-mod user_approval_widget;
 mod version;
 mod wrapping;
