--- codex-rs/tui/src/user_approval_widget.rs
@@ -1,15 +1,0 @@
-use std::path::PathBuf;
-
-/// Request coming from the agent that needs user approval.
-pub(crate) enum ApprovalRequest {
-    Exec {
-        id: String,
-        command: Vec<String>,
-        reason: Option<String>,
-    },
-    ApplyPatch {
-        id: String,
-        reason: Option<String>,
-        grant_root: Option<PathBuf>,
-    },
-}
